# Conflicts:
#	src/main/java/uz/sfera/edu_platform/service/LessonService.java
#	src/main/java/uz/sfera/edu_platform/service/QuizService.java

+++ src/main/java/uz/sfera/edu_platform/entity/Result.java
@@ -23,7 +23,7 @@ public class Result {
     private int totalQuestion;
 
     @Column(nullable = false)
-    private int correctAnswers;
+    private long correctAnswers;
 
     @Column(nullable = false)
     private Long timeTaken;
